+++ packages/starrating/index.js
@@ -1,0 +1,4 @@
+const css = require('./css')
+const react = require('./react')
+
+module.exports = { css: css, react: react }