+++ packages/checkbox/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-checkbox@0.3.11...@pluralsight/ps-design-system-checkbox@0.3.12) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-checkbox
+
+
+
+
+
+## [0.3.11](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-checkbox@0.3.10...@pluralsight/ps-design-system-checkbox@0.3.11) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-checkbox
+
+
+
+
+
 ## [0.3.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-checkbox@0.3.9...@pluralsight/ps-design-system-checkbox@0.3.10) (2018-11-09)