+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@1.5.6...@pluralsight/ps-design-system-drawer@1.5.7) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
+## [1.5.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@1.5.5...@pluralsight/ps-design-system-drawer@1.5.6) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
 ## [1.5.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@1.5.4...@pluralsight/ps-design-system-drawer@1.5.5) (2018-11-09)