+++ packages/icon/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.13.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-icon@8.13.4...@pluralsight/ps-design-system-icon@8.13.5) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-icon
+
+
+
+
+
+## [8.13.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-icon@8.13.3...@pluralsight/ps-design-system-icon@8.13.4) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-icon
+
+
+
+
+
 ## [8.13.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-icon@8.13.2...@pluralsight/ps-design-system-icon@8.13.3) (2018-11-09)