+++ packages/layout/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-layout@3.0.6...@pluralsight/ps-design-system-layout@3.0.7) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-layout
+
+
+
+
+
+## [3.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-layout@3.0.5...@pluralsight/ps-design-system-layout@3.0.6) (2018-11-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-layout
+
+
+
+
+
 ## [3.0.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-layout@3.0.4...@pluralsight/ps-design-system-layout@3.0.5) (2018-11-09)